Made An Update Context Pack in Database Test and Add Function
Created a test to test if the requests to update the context pack is received. It then tests is it updates the database. An add function is created and used in the update database test. It applies the POST method onto the context pack and adds it to the database.