Create A Route Method and Set Data method
Created a route method that takes the user to add wordlist after clicking on the button. Created a set data button which is used in the router. Sends the context pack over to the add wordlist. This context pack is used as the default context pack, which can be changed.